get returns option for dba/el/file

+++ src/main/scala/rng/HashingExtension.scala
@@ -9,7 +9,7 @@ import scala.collection.{SortedMap}
 class HashingImpl(config: Config) extends Extension {
   val hashLen = config.getInt("hash-length")
   val bucketsNum = config.getInt("buckets")
-  val bucketRange = (math.pow(2, hashLen) / bucketsNum).ceil.toInt
+  val bucketRange = (math.pow(2, hashLen.toDouble) / bucketsNum).ceil.toInt
 
   def hash(word: Array[Byte]): Int = {
     implicit val digester = MessageDigest.getInstance("MD5")
